[6502 alu] Setting N before an SBC operation is wrong I think
I've always set this but I can't write a unit test to pick it up. I
don't thin the ALU is responsible for setting the N flag, that's the
CPU when a value is loaded somewhere.  Meaning that the ALU can't be
tested for it.  Removed it for now until I can research it further.