Biting the bullet and moving to custom Byte representation from ints.
This isn't the prettiest syntax, I've did the best I could to make it nicer with custom equals() but at least the concept of bytes is abstracted and we can change the implementation easier.  I'll think of better syntax as I go.
This was also a good exercise in refactoring a large, highly unit tested piece of code.