Cleaning up unecessary code
OpCode already knows what it's AddressingMode and Operation are, so
explicitly stating them in a passed in function for each was overkill.
All I needed to do was call whatever AddressingMode and Operation was
part of that OpCode.  This is the same for all OpCodes so one method
for all replaces one for each.