Added BRK to functional enum pattern and restructured & added missing operation Javadoc
Now that all opcodes than can be enum-ed, have been, we can get rid of
the specific cases and apply the default to all but the exceptions.
Would still be nice to come up with a JMP strategy that allows us to
get rid of the case entirely but for now, it's much neater.