[exp/serialization] Adding 6502 serialization and deserializtion
I'm not sure if this is even the route I want to go down.  6502 doesn't
actually have any state by itself, only its registers and memory.
Memory isn't the 6502 and -at least in our case- neither are the
Registers.  This is just an experiment into serialization now so we can
think about it later.